Changed physics to include momentum forces from the lever on the ball. Fov 60 -> 65. Added post processing.
Using the rotate methods, levers didn't add force to the to the ball from their rotational inertia . To change this, torque is added to the levers instead so the forces are included in the physics calculation. This adds a more realistic feel to the game and makes it harder. The former physics are still accessible by checking the new easy mode unlockable.  The main camera's field of view was also increased from 60 to 65.